+++ tests/dummy/app/templates/snippets/position-1-js.js
@@ -3,5 +3,6 @@ import Controller from '@ember/controller';
 export default Controller.extend({
   horizontalPosition: 'auto',
   verticalPosition: 'auto',
-  buttonPosition: 'left'
+  buttonPosition: 'left',
+  renderInPlace: false
 });