+++ tests/integration/components/basic-dropdown-test.js
@@ -3,8 +3,8 @@ import { registerDeprecationHandler } from '@ember/debug';
 import { module, test } from 'qunit';
 import { setupRenderingTest, render } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
-import { clickTrigger } from '../../helpers/ember-basic-dropdown';
-import { click, find, focus } from 'ember-native-dom-helpers';
+import { clickTrigger } from 'ember-basic-dropdown/test-support/helpers';
+import { click, find, focus, triggerEvent } from 'ember-native-dom-helpers';
 
 let deprecations = [];
 
@@ -34,7 +34,24 @@ module('Integration | Component | basic-dropdown', function(hooks) {
     await clickTrigger();
     assert.ok(find('#dropdown-is-opened'), 'The dropdown is opened');
     await clickTrigger();
-    assert.notOk(find('#dropdown-is-opened'), 'The dropdown is again');
+    assert.notOk(find('#dropdown-is-opened'), 'The dropdown is closed again');
+  });
+
+  test('The mousedown event with the right button doesn\'t open it', async function(assert) {
+    assert.expect(2);
+
+    await render(hbs`
+      {{#basic-dropdown as |dd|}}
+        {{#dd.trigger}}Click me{{/dd.trigger}}
+        {{#if dd.isOpen}}
+          <div id="dropdown-is-opened"></div>
+        {{/if}}
+      {{/basic-dropdown}}
+    `);
+
+    assert.notOk(find('#dropdown-is-opened'), 'The dropdown is closed');
+    await triggerEvent('.ember-basic-dropdown-trigger', 'mousedown', { button: 2 });
+    assert.notOk(find('#dropdown-is-opened'), 'The dropdown is closed');
   });
 
   test('Its `open` action opens the dropdown', async function(assert) {
@@ -348,6 +365,36 @@ module('Integration | Component | basic-dropdown', function(hooks) {
 
     await clickTrigger();
     assert.ok(find('.ember-basic-dropdown'), 'The trigger has a special `--in-place` class');
+  });
+
+  test('When rendered in-place, it prefers right over left with position "auto-right"', async function(assert) {
+    assert.expect(2);
+
+    await render(hbs`
+      {{#basic-dropdown renderInPlace=true horizontalPosition="auto-right" as |dropdown|}}
+        {{#dropdown.trigger}}Press me{{/dropdown.trigger}}
+        {{#dropdown.content}}<h3>Content of the dropdown</h3>{{/dropdown.content}}
+      {{/basic-dropdown}}
+    `);
+
+    await clickTrigger();
+    assert.ok(find('.ember-basic-dropdown-trigger').classList.contains('ember-basic-dropdown-trigger--right'), 'The proper class has been added');
+    assert.ok(find('.ember-basic-dropdown-content').classList.contains('ember-basic-dropdown-content--right'), 'The proper class has been added');
+  });
+
+  test('When rendered in-place, it applies right class for position "right"', async function(assert) {
+    assert.expect(2);
+
+    await render(hbs`
+      {{#basic-dropdown renderInPlace=true horizontalPosition="right" as |dropdown|}}
+        {{#dropdown.trigger}}Press me{{/dropdown.trigger}}
+        {{#dropdown.content}}<h3>Content of the dropdown</h3>{{/dropdown.content}}
+      {{/basic-dropdown}}
+    `);
+
+    await clickTrigger();
+    assert.ok(find('.ember-basic-dropdown-trigger').classList.contains('ember-basic-dropdown-trigger--right'), 'The proper class has been added');
+    assert.ok(find('.ember-basic-dropdown-content').classList.contains('ember-basic-dropdown-content--right'), 'The proper class has been added');
   });
 
   test('[ISSUE #127] Having more than one dropdown with `renderInPlace=true` raises an exception', async function(assert) {