+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 
+- [BREAKING] Delete the `/test-support` folder, use `addon-test-support` instead. That means people
+  should import helpers from `ember-basic-dropdown/test-support/helpers` instead of using relative paths
+  like `../../helpers/ember-basic-dropdown`, as they are brittle and change with nesting.
+- [BUGFIX] Ensure that the dropdown is not open by the right button of the mouse.
+
+# 0.33.10
+- [ENHANCEMENT] Allow `horizontalPosition` to work when `renderInPlace=true`
+
 # 0.33.9
 - [ENHANCEMENT] Allow to bind the role attribute of the trigger
 - [ENHANCEMENT] Added `preventScroll=true` option to the `dropdown.content` component to "frezze" all